3 fix installed packages

+++ configs/310_alias.zsh
@@ -1,35 +1,42 @@
+#!/bin/zsh
+
 # list of alias
 # - grep: ggrep or grep with color
 # - dc: docker compose, and some alias for docker compose command
 # - ls: exa if available
 #   - tree: exa --tree
-# - cat: bat if available
+
+function __print_alias() {
+  printf "        alias: \e[36m$1\e[m = \e[36m$2\e[m\n"
+}
 
 # grep
 if type ggrep > /dev/null; then
   alias grep='ggrep --color'
+  __print_alias "grep" "ggrep"
 else
   alias grep='grep --color=always'
 fi
 
 # docker compose
-alias docker-compose='docker compose'
-alias dc='docker compose'
-alias dc-build='docker compose build'
-alias dc-run='docker compose run'
-alias dc-up='docker compose up'
-alias dc-down='docker compose down'
-alias dc-logs='docker compose logs'
-alias dc-ps='docker compose ps'
-alias dc-exec='docker compose exec'
+if type docker-compose > /dev/null; then
+  alias docker-compose='docker compose'
+  alias dc='docker compose'
+  alias dc-build='docker compose build'
+  alias dc-run='docker compose run'
+  alias dc-up='docker compose up'
+  alias dc-down='docker compose down'
+  alias dc-logs='docker compose logs'
+  alias dc-ps='docker compose ps'
+  alias dc-exec='docker compose exec'
+  __print_alias "dc" "docker-compose"
+fi
 
 # ls -> exa 
-printf "\e[36mexa\e[m:"
 if type exa > /dev/null; then
-  printf " \e[32mok\e[m.\n"
   alias ls='exa --sort=extension --group-directories-first --group'
+  __print_alias "ls" "exa"
 else
-  printf "not available.\n"
   if [ "$(uname)" "==" "Darwin" ]; then
     alias ls='ls -G'
   else
@@ -39,18 +46,8 @@ fi
 alias ll='ls -al'
 alias la='ls -a'
 
-# cat -> bat
-printf "\e[36mbat\e[m:"
-if type bat > /dev/null; then
-  printf " \e[32mok\e[m.\n"
-  alias cat='bat --style="plain" --theme="ansi"'
-else
-  printf "not available.\n"
-fi
-
 # tree
 if type exa > /dev/null; then
   alias tree='exa --tree'
-else
-  echo > /dev/null
+  __print_alias "tree" "exa --tree"
 fi
